requirements.txt
alabaster==0.7.12
alembic==1.4.3
entrypoints==0.3
flake8-import-order==0.18.1
Flask==1.0.2
Flask-Bcrypt==0.7.1
Flask-Cors==3.0.9
Flask-Migrate==2.5.3
flask-restplus==0.13.0
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.4

grpcio==1.32.0
grpcio-tools==1.32.0
h5py==2.9.0
html5lib==1.0.1
imageio==2.5.0
imagesize==1.1.0
importlib-resources==3.0.0
imutils==0.5.3
Jinja2==2.10
json5==0.9.5
jsonschema==3.0.1
Keras==2.2.4
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
lxml==4.3.2
Mako==1.1.3
Markdown==3.2.2
MarkupSafe==1.1.1
matplotlib==3.0.3
numpy==1.15.4
numpydoc==0.8.0
oauthlib==3.1.0
opencv-contrib-python==4.4.0.42
opencv-python==4.0.0.21
cloudpickle==0.8.0
pathlib2==2.3.3
pickleshare==0.7.5
Pillow==6.2.2
progress==1.5
prometheus-client==0.6.0
promise==2.3
prompt-toolkit==3.0.7
protobuf==3.13.0
psycopg2-binary==2.8.6
PyJWT==1.7.1
seaborn==0.9.0
SecretStorage==3.1.1
six==1.15.0
sklearn==0.0
statsmodels==0.9.0
tblib==1.3.2
tensorboard==1.13.1
tensorboard-plugin-wit==1.7.0
tensorboardX==2.1
tensorflow==1.13.1
tensorflow-datasets==4.0.1
tensorflow-estimator==1.13.0
Werkzeug==0.14.1
xlrd==1.2.0
XlsxWriter==1.1.5